requirements.txt
altgraph==0.17
asgiref==3.2.7
astroid==2.3.3
backoff==1.10.0
bcrypt==3.1.7
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
colorama==0.4.3
cryptography==2.9.2
future==0.18.2
idna==2.9
isort==4.3.21
lazy-object-proxy==1.4.3
lxml==4.5.0
mccabe==0.6.1
opencage==1.2.1
paramiko==2.7.1
pefile==2019.4.18
ping3==2.6.5
prettytable==0.7.2
pycparser==2.20
PyInstaller==3.6
pylint==2.4.4
PyNaCl==1.4.0
pyOpenSSL==19.1.0
pyperclip==1.8.0
pyping==0.0.5
python-dotenv==0.13.0
pythonping==1.0.10
pytz==2020.1
pywin32-ctypes==0.2.0
requests==2.23.0
RouterOS-api==0.17.0
selenium==3.141.0
six==1.14.0
soupsieve==2.0
sqlparse==0.3.1
urllib3==1.25.9
wrapt==1.11.2
